requirements.txt
# Core Django
Django>=5.0
psycopg[binary]>=3.1


# Environment variables
python-decouple==3.8

# API requests
requests==2.31.0

# Heroku deployment
gunicorn==21.2.0
whitenoise==6.6.0
dj-database-url==2.1.0

# Development
python-dotenv==1.0.0

# Caching
django-redis==5.4.0

# Forms and widgets
django-crispy-forms>=2.3
crispy-bootstrap5>=2025.6



# Timezone support
pytz==2024.1